Add files via upload
Taking care of one's health is an important part of most people's lives. Not everyone has the opportunity to undergo regular professional medical examinations. However, there are a number of simple tests that can help you accurately assess your health. One such test is the Rufier test, which allows you to determine the condition of the body's cardiorespiratory system. 

The Rufier test helps users assess the state of their body's cardiovascular system.

It is designed to monitor cardiac performance during physical exertion and evaluate the effectiveness of that performance.